[Controls] Reduce flakiness of options list tests (#262638)
Closes https://github.com/elastic/kibana/issues/246769

## Summary

This PR reduces the flakiness of all the options list control tests
**and** improves their speed by:

1. Relying on the ES + Kibana archives where possible
2. Separating out the long `options_list_dashboard_interaction.ts` test
file into multiple smaller test suites

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [x] [Flaky Test
Runner](https://ci-stats.kibana.dev/trigger_flaky_test_runner/1) was
used on any tests changed
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)
- [x] Review the [backport
guidelines](https://docs.google.com/document/d/1VyN5k91e5OVumlc0Gb9RPa3h1ewuPE705nRtioPiTvY/edit?usp=sharing)
and apply applicable `backport:*` labels.

(cherry picked from commit 101f87bc3944710d99bf4ea854b9de20949bd6b7)
