+++ src/platform/test/functional/apps/dashboard_elements/controls/options_list/options_list_dashboard_interaction.ts
@@ -9,60 +9,32 @@
 
 import { pick } from 'lodash';
 
-import { OPTIONS_LIST_CONTROL } from '@kbn/controls-constants';
 import expect from '@kbn/expect';
 
 import { OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS } from '../../../../page_objects/dashboard_page_controls';
 import type { FtrProviderContext } from '../../../../ftr_provider_context';
-import { OPTIONS_LIST_DASHBOARD_NAME } from '.';
+
+const OPTIONS_LIST_DASHBOARD_NAME = 'Test Options List Control';
 
 export default function ({ getService, getPageObjects }: FtrProviderContext) {
   const queryBar = getService('queryBar');
   const pieChart = getService('pieChart');
-  const elasticChart = getService('elasticChart');
   const filterBar = getService('filterBar');
-  const testSubjects = getService('testSubjects');
-  const dashboardAddPanel = getService('dashboardAddPanel');
-  const dashboardPanelActions = getService('dashboardPanelActions');
 
-  const { dashboardControls, timePicker, console, common, dashboard, header, settings } =
-    getPageObjects([
-      'dashboardControls',
-      'timePicker',
-      'dashboard',
-      'settings',
-      'console',
-      'common',
-      'header',
-    ]);
+  const { dashboardControls, timePicker, dashboard, header } = getPageObjects([
+    'dashboardControls',
+    'timePicker',
+    'dashboard',
+    'header',
+  ]);
 
   describe('Interactions between options list and dashboard', () => {
     let controlId: string;
 
-    const returnToDashboard = async () => {
-      await dashboard.navigateToApp();
-      await header.waitUntilLoadingHasFinished();
-      await elasticChart.setNewChartUiDebugFlag();
-      await dashboard.loadDashboardInEditMode(OPTIONS_LIST_DASHBOARD_NAME);
-    };
-
     before(async () => {
-      await dashboard.ensureDashboardIsInEditMode();
-      await dashboardControls.createControl({
-        controlType: OPTIONS_LIST_CONTROL,
-        dataViewTitle: 'animals-*',
-        fieldName: 'sound.keyword',
-      });
+      await dashboard.loadDashboardInEditMode(OPTIONS_LIST_DASHBOARD_NAME);
       controlId = (await dashboardControls.getAllControlIds())[0];
-      await dashboardAddPanel.addVisualization('Rendering-Test:-animal-sounds-pie');
-      await dashboard.clickQuickSave();
       await header.waitUntilLoadingHasFinished();
-    });
-
-    after(async () => {
-      await dashboardControls.deleteAllPinnedControls();
-      await dashboardPanelActions.removePanelByTitle('Rendering Test: animal sounds pie');
-      await dashboard.clickQuickSave();
     });
 
     describe('Applies query settings to controls', () => {
@@ -159,25 +131,14 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
             invalidSelections: [],
           });
         });
+      });
 
-        after(async () => {
-          await filterBar.removeAllFilters();
-        });
+      after(async () => {
+        await dashboard.clickDiscardChanges();
       });
     });
 
-    // FLAKY: https://github.com/elastic/kibana/issues/239769
-    describe.skip('Selections made in control apply to dashboard', () => {
-      it('Shows available options in options list', async () => {
-        await queryBar.setQuery('');
-        await queryBar.submitQuery();
-        await header.waitUntilLoadingHasFinished();
-        await dashboardControls.ensureAvailableOptionsEqual(controlId, {
-          suggestions: OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS,
-          invalidSelections: [],
-        });
-      });
-
+    describe('Selections made in control apply to dashboard', () => {
       it('Can search options list for available options', async () => {
         await dashboardControls.optionsListOpenPopover(controlId);
         await dashboardControls.optionsListPopoverSearchForOption('meo');
@@ -237,16 +198,12 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
       });
 
       it('excluding selections has expected results', async () => {
-        await dashboard.clickQuickSave();
-        await header.waitUntilLoadingHasFinished();
-
         await dashboardControls.optionsListOpenPopover(controlId);
         await dashboardControls.optionsListPopoverSetIncludeSelections(false);
         await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
 
         await dashboard.waitForRenderComplete();
         expect(await pieChart.getPieSliceCount()).to.be(5);
-        await dashboard.clearUnsavedChanges();
       });
 
       it('including selections has expected results', async () => {
@@ -256,172 +213,43 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         await dashboard.waitForRenderComplete();
         expect(await pieChart.getPieSliceCount()).to.be(2);
-        await dashboard.clearUnsavedChanges();
-      });
-
-      describe('discarding changes', () => {
-        describe('changes can be discarded', () => {
-          let selections = '';
-
-          beforeEach(async () => {
-            await dashboardControls.optionsListOpenPopover(controlId);
-            await dashboardControls.optionsListPopoverSelectOption('bark');
-            await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
-            selections = await dashboardControls.optionsListGetSelectionsString(controlId);
-            expect(selections).to.equal('hiss, grr, bark');
-          });
-
-          afterEach(async () => {
-            selections = await dashboardControls.optionsListGetSelectionsString(controlId);
-            expect(selections).to.equal('hiss, grr');
-          });
-
-          it('by clicking the discard changes button', async () => {
-            await dashboard.clickDiscardChanges();
-          });
-
-          it('by switching to view mode', async () => {
-            await dashboard.clickCancelOutOfEditMode();
-          });
-        });
-
-        it('dashboard does not load with unsaved changes when changes are discarded', async () => {
-          await dashboard.switchToEditMode();
-          await dashboard.ensureMissingUnsavedChangesNotification();
-        });
-      });
-    });
-
-    describe('Test data view runtime field', () => {
-      const FIELD_NAME = 'testRuntimeField';
-      const FIELD_VALUES = {
-        G:
-          OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.growl +
-          OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.grr +
-          OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.grrr,
-        H: OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.hiss,
-        B:
-          OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.bark +
-          OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS['bow ow ow'],
-        R: OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.ruff,
-        M: OPTIONS_LIST_ANIMAL_SOUND_SUGGESTIONS.meow,
-      };
-
-      before(async () => {
-        await common.navigateToApp('settings');
-        await settings.clickKibanaIndexPatterns();
-        await settings.clickIndexPatternByName('animals-*');
-        await settings.addRuntimeField(
-          FIELD_NAME,
-          'keyword',
-          `emit(doc['sound.keyword'].value.substring(0, 1).toUpperCase())`
-        );
-        await returnToDashboard();
-        await dashboardControls.deleteAllPinnedControls();
-      });
-
-      it('can create options list control on runtime field', async () => {
-        await dashboardControls.createControl({
-          controlType: OPTIONS_LIST_CONTROL,
-          fieldName: FIELD_NAME,
-          dataViewTitle: 'animals-*',
-        });
-        expect(await dashboardControls.getControlsCount()).to.be(1);
-      });
-
-      it('new control has expected suggestions', async () => {
-        controlId = (await dashboardControls.getAllControlIds())[0];
-        await dashboardControls.ensureAvailableOptionsEqual(controlId, {
-          suggestions: FIELD_VALUES,
-          invalidSelections: [],
-        });
-      });
-
-      it('making selection has expected results', async () => {
-        await dashboardControls.optionsListOpenPopover(controlId);
-        await dashboardControls.optionsListPopoverSelectOption('B');
-        await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
-
-        await dashboard.waitForRenderComplete();
-        expect(await pieChart.getPieChartLabels()).to.eql(['bark', 'bow ow ow']);
       });
 
       after(async () => {
-        await dashboardControls.deleteAllPinnedControls();
-        await dashboard.clickQuickSave();
-        await header.waitUntilLoadingHasFinished();
-
-        await common.navigateToApp('settings');
-        await settings.clickKibanaIndexPatterns();
-        await settings.clickIndexPatternByName('animals-*');
-        await settings.filterField('testRuntimeField');
-        await testSubjects.click('deleteField');
-        await settings.confirmDelete();
+        await dashboard.clickDiscardChanges();
       });
     });
 
-    describe('Test exists query', () => {
-      const newDocuments: Array<{ index: string; id: string }> = [];
+    describe('discarding changes', () => {
+      describe('changes can be discarded', () => {
+        let selections = '';
 
-      const addDocument = async (index: string, document: string) => {
-        await console.enterText('\nPOST ' + index + '/_doc/\n{\n ' + document + '\n}');
-        await console.clickPlay();
-        await header.waitUntilLoadingHasFinished();
-        const response = JSON.parse(await console.getOutputText());
-        newDocuments.push({ index, id: response._id });
-      };
+        beforeEach(async () => {
+          await dashboardControls.optionsListOpenPopover(controlId);
+          await dashboardControls.optionsListPopoverSelectOption('bark');
 
-      before(async () => {
-        await common.navigateToApp('console');
-        await console.skipTourIfExists();
-        await console.clearEditorText();
-        await addDocument(
-          'animals-cats-2018-01-01',
-          '"@timestamp": "2018-01-01T16:00:00.000Z", \n"name": "Rosie", \n"sound": "hiss"'
-        );
-        await returnToDashboard();
-
-        await dashboardControls.createControl({
-          controlType: OPTIONS_LIST_CONTROL,
-          dataViewTitle: 'animals-*',
-          fieldName: 'animal.keyword',
-          title: 'Animal',
+          await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
+          selections = await dashboardControls.optionsListGetSelectionsString(controlId);
+          expect(selections).to.equal('bark');
         });
-        controlId = (await dashboardControls.getAllControlIds())[0];
-        await header.waitUntilLoadingHasFinished();
+
+        afterEach(async () => {
+          selections = await dashboardControls.optionsListGetSelectionsString(controlId);
+          expect(selections).to.equal('Any');
+        });
+
+        it('by clicking the discard changes button', async () => {
+          await dashboard.clickDiscardChanges();
+        });
+
+        it('by switching to view mode', async () => {
+          await dashboard.clickCancelOutOfEditMode();
+        });
       });
 
-      it('creating exists query has expected results', async () => {
-        await dashboard.waitForRenderComplete();
-        expect((await pieChart.getPieChartValues())[0]).to.be(6);
-        await dashboardControls.optionsListOpenPopover(controlId);
-        await dashboardControls.optionsListPopoverSelectExists();
-        await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
-        await dashboard.waitForRenderComplete();
-
-        expect(await pieChart.getPieSliceCount()).to.be(5);
-        expect((await pieChart.getPieChartValues())[0]).to.be(5);
-      });
-
-      it('negating exists query has expected results', async () => {
-        await dashboardControls.optionsListOpenPopover(controlId);
-        await dashboardControls.optionsListPopoverSetIncludeSelections(false);
-        await dashboardControls.optionsListEnsurePopoverIsClosed(controlId);
-        await dashboard.waitForRenderComplete();
-
-        expect(await pieChart.getPieSliceCount()).to.be(1);
-        expect((await pieChart.getPieChartValues())[0]).to.be(1);
-      });
-
-      after(async () => {
-        await common.navigateToApp('console');
-        await console.clearEditorText();
-        for (const { index, id } of newDocuments) {
-          await console.enterText(`\nDELETE /${index}/_doc/${id}`);
-          await console.clickPlay();
-          await header.waitUntilLoadingHasFinished();
-        }
-        await returnToDashboard();
+      it('dashboard does not load with unsaved changes when changes are discarded', async () => {
+        await dashboard.loadDashboardInEditMode(OPTIONS_LIST_DASHBOARD_NAME);
+        await dashboard.ensureMissingUnsavedChangesNotification();
       });
     });
   });
