+++ src/platform/test/functional/fixtures/es_archiver/dashboard_elements/controls/chaining/data.json
