+++ src/platform/test/functional/fixtures/es_archiver/dashboard_elements/controls/test_exists/data.json
@@ -1,0 +1,12 @@
+{
+  "type": "doc",
+  "value": {
+    "id": "1",
+    "index": "animals-cats-2018-01-01",
+    "source": {
+      "@timestamp": "2018-01-01T16:00:00.000Z",
+      "name": "Rosie",
+      "sound": "hiss"
+    }
+  }
+}
